:recycle: refactor: filters 파일을 만들어 class 분리
